refactor: 글로벌모달 렌더링 로직 최적화
